refactor(scripts): switching to use hdwallet for all scripts
Removed web3 script and added all web3 functionality to hdwallet.
Environemnt variable for auction changed to AUCTION_ADDRESS.
Environment variable for MNEMONIC must now be set using local ganache generated mnemonic.
